miner: use maskHash for RPC `getwork` key
HSD Full Node RPC module tracks mining work by maskHash, not
merkleRoot. With this change, the miner will not be stopped
and reset every poll, unless the work has actually changed.

We also switch to comparing buffers instead of strings, which
is slightly faster in nodejs. Strings are only needed for JSON
output, and in submitting JSON-RPC.