opencl: adjust threads and blocks options.
The default threads (total work items) and blocks (work group size) cli
options have the potential to cause errors for opencl capable devices.
This commit makes sure the options are lowered to the device's default
values, if necessary. If the resultant total work items value is not
divisible by the work group size, the total work items value is lowered
to the closest multiple of the work group size. We also print the values
to alert users of any changes to cli options.